requirements.txt
altair==5.5.0
<<<<<<< HEAD
=======
annotated-doc==0.0.4
>>>>>>> be78c6f8d0aba9102694f8c7504706466ce8a2fb
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
<<<<<<< HEAD
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distlib==0.4.0
filelock==3.18.0
Flask==3.1.2
fonttools==4.58.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
google-genai==1.52.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==2.12.0
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==11.2.1
platformdirs==4.3.8
protobuf==6.33.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pygame==2.6.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.3
rpds-py==0.29.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
=======
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.123.0
fonttools==4.61.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
narwhals==2.12.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.1
pyarrow==21.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
starlette==0.50.0
>>>>>>> be78c6f8d0aba9102694f8c7504706466ce8a2fb
streamlit==1.51.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
<<<<<<< HEAD
urllib3==2.4.0
virtualenv==20.33.1
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
pandas
numpy
matplotlib
seaborn
python-dotenv

=======
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
>>>>>>> be78c6f8d0aba9102694f8c7504706466ce8a2fb
